+++ packages/wallet/src/index.ts
@@ -1,8 +1,9 @@
 export { Wallet } from './Wallet';
-export type { WalletOptions } from './types';
 export type {
   DefaultActions,
   DefaultEvents,
   RootMessenger,
   WalletDestroyedEvent,
 } from './initialization';
+export type { WalletOptions } from './types';
+export { importSecretRecoveryPhrase } from './utilities';